% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kber.R
\name{k_AIC}
\alias{k_AIC}
\title{Calculate the AIC values of a series of models, varying k}
\usage{
k_AIC(di, i, k, response, family, weights = NULL)
}
\arguments{
\item{di}{a vector of stimulus distances (see \code{\link[=stimulus_contrast]{stimulus_contrast()}})}

\item{i}{a vector of stimulus magnitudes (see \code{\link[=stimulus_contrast]{stimulus_contrast()}})}

\item{k}{a constant which determines the strength of the magnitude effect (see \code{\link[=stimulus_contrast]{stimulus_contrast()}})}

\item{response}{the response variable of the model (discrimination between stimuli)}

\item{family}{model family to use for \code{\link[=stimulus_model]{stimulus_model()}} (see \code{\link[=glm]{glm()}})}

\item{weights}{weights of each observation, used in binomial models (see \code{\link[=glm]{glm()}}).
Leave empty if each observation is 1 decision (binary stimuli)
Solid black line shows how the AIC of \code{\link[=stimulus_model]{stimulus_model()}} varies with \code{k};
red vertical line shows the best fitting value of \code{k};
dashed horizontal line crosses the AIC curve at the 95\% confidence intervals for this \code{k} estimate.}
}
\value{
A summary table containing the best fitting value of \code{k} which minimises the model AIC,
its 95\% confidence intervals (\code{lower_95} and \code{upper_95}),
and the AIC values of \code{\link[=stimulus_model]{stimulus_model()}} for these values of k.
}
\description{
Fits a generalised linear models with \code{\link[=stimulus_model]{stimulus_model()}} for a given value of \code{k} (see \code{\link[=stimulus_contrast]{stimulus_contrast()}}),
and computes the AIC of the extracting model. This is mostly used by other functions,
the only reason you'd really need it is if you want to make custom diagnostic plots with \code{ggplot}.
}
\examples{
data(stimuli)
#data contains columns for two stimuli of different magnitudes, a and b

stimuli$abs_diff <- abs(stimuli$stimulus_a - stimuli$stimulus_b)
#stimulus distance = absolute difference between a and b

stimuli$mean_ab <- (stimuli$stimulus_a + stimuli$stimulus_b)/2
#stimulus magnitude = mean magnitude of a and b

#get the AIC of stimulus_model for k = 1
k_AIC(di = stimuli$abs_diff, i = stimuli$mean_ab, k = 1,
response = stimuli$discrimination, family = binomial(link="logit"))
}
