% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kber.R
\docType{data}
\name{stimuli}
\alias{stimuli}
\title{Stimulus discrimination data}
\format{
A data frame with 119 rows and 3 variables:
\describe{
\item{stimulus_a}{magnitude of stimulus A}
\item{stimulus_b}{magnitude of stimulus B}
\item{discrimination}{whether the subject discriminated between A and B (0 = no, 1 = yes)}
}
}
\source{
Dixit, Tanmay; Apostol, Andrei L.; Chen, Kuan-Chi; Fulford,
Anthony J. C.; Town, Christopher P.; Spottiswoode, Claire N. (2022).
Supplementary material from "Visual complexity of egg patterns predicts
egg rejection according to Weber's law". The Royal Society. Collection.
https://doi.org/10.6084/m9.figshare.c.6066380.v2
}
\usage{
stimuli
}
\description{
A dataset containing discrimination performance between stimuli,
originally taken from prinias discriminating between eggs of different
complexities (Dixit et al. 2022)
}
\keyword{datasets}
