% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kber.R
\name{stimulus_contrast}
\alias{stimulus_contrast}
\title{Stimulus contrast}
\usage{
stimulus_contrast(di, i, k)
}
\arguments{
\item{di}{a vector of stimulus distances, where stimulus distance is the absolute difference between the magnitudes of two stimuli \code{a} and \code{b}.}

\item{i}{a vector of stimulus magnitudes, where stimulus magnitude is usually the mean magnitude of two stimuli \code{a} and \code{b}
(but other measures could be used).}

\item{k}{a constant which determines the strength of the magnitude effect.
When \code{k = 0}, only stimulus distance is used to predict discrimination;
when \code{k = 1}, discrimination follows Weber's law;
\verb{0 < k < 1} is a near miss to Weber's law;
\code{k > 1} is an opposite miss to Weber's law.}
}
\value{
The contrast between two stimuli, computed as \code{di/(i^k)}.
}
\description{
Computes the stimulus contrast, defined as \code{distance/(magnitude^k)}, for a given value of \code{k}.
}
\examples{
data(stimuli)
#data contains columns for two stimuli of different magnitudes, a and b
stimuli$abs_diff <- abs(stimuli$stimulus_a - stimuli$stimulus_b) #stimulus distance = absolute difference between a and b
stimuli$mean_ab <- (stimuli$stimulus_a + stimuli$stimulus_b)/2 #stimulus magnitude = mean magnitude of a and b
#get stimulus contrast for k=1.2
stimuli$contrast <- stimulus_contrast(di=stimuli$abs_diff, i=stimuli$mean_ab, k=1.2);
}
