% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kber.R
\name{stimulus_model}
\alias{stimulus_model}
\title{Stimulus model}
\usage{
stimulus_model(di, i, k, response, family, weights = NULL)
}
\arguments{
\item{di}{a vector of stimulus distances (see \code{\link[=stimulus_contrast]{stimulus_contrast()}})}

\item{i}{a vector of stimulus magnitudes (see \code{\link[=stimulus_contrast]{stimulus_contrast()}})}

\item{k}{a constant which determines the strength of the magnitude effect (see \code{\link[=stimulus_contrast]{stimulus_contrast()}})}

\item{response}{the response variable of the model (discrimination between stimuli)}

\item{family}{model family (see \code{\link[=glm]{glm()}})}

\item{weights}{weights of each observation, used in binomial models (see \code{\link[=glm]{glm()}}).
Leave empty if each observation is 1 decision (binary stimuli)}
}
\value{
A fitted generalised linear model of the form \verb{glm(response ~ }
\code{\link[=stimulus_contrast]{stimulus contrast(di, i, k)}}\verb{, family=family, weights=weights)}
}
\description{
Fits a generalised linear model, with \code{\link[=stimulus_contrast]{stimulus contrast(di, i, k)}}
as the predictor variable and \code{response} as the response variable
}
\examples{
data(stimuli)
#data contains columns for two stimuli of different magnitudes, a and b

stimuli$abs_diff <- abs(stimuli$stimulus_a - stimuli$stimulus_b)
#stimulus distance = absolute difference between a and b

stimuli$mean_ab <- (stimuli$stimulus_a + stimuli$stimulus_b)/2
#stimulus magnitude = mean magnitude of a and b

#get stimulus model for k=1
 stimulus_model(di = stimuli$abs_diff, i = stimuli$mean_ab, k = 1,
 response = stimuli$discrimination, family = binomial(link = "logit"))
}
